requirements.txt
# Saudi Stock Bot Requirements
# Specific versions to avoid compatibility issues
Flask==2.0.1
Werkzeug==2.0.1
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
yfinance==0.2.28
SQLAlchemy==2.0.20
pymongo==4.5.0
nltk==3.8.1
transformers==4.34.0
scikit-learn==1.3.0
beautifulsoup4==4.12.2
lxml==4.9.3
gunicorn==21.2.0
pytest==7.4.0
python-dateutil==2.8.2
pytz==2023.3
ta==0.10.2  # Technical Analysis library
googletrans==4.0.0-rc1  # For translating Arabic news
discord-webhook==1.3.0  # For Discord notifications
